Added registration and login functionalities

+++ project-code/src/authentication/registration/UserRegistration.java
@@ -1,8 +1,10 @@
 package authentication.registration;
 
 import authentication.SecurityQuestion;
-import authentication.utils.*;
+import authentication.utils.HashingService;
 
+import java.io.BufferedReader;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -21,6 +23,11 @@ public class UserRegistration {
         String hashedUsername = HashingService.hashText(username);
         String hashedPassword = HashingService.hashText(password);
 
+        if (checkUserExists(hashedUsername)) {
+            System.out.println("Username taken");
+            return;
+        }
+
         try (FileWriter fileWriter = new FileWriter("User_Profile.txt", true)) {
             StringBuilder builder = new StringBuilder();
 
@@ -37,6 +44,23 @@ public class UserRegistration {
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+
+    private boolean checkUserExists(String username) {
+        try (BufferedReader bufferedReader = new BufferedReader(new FileReader("User_Profile.txt"))) {
+            String line;
+            while ((line = bufferedReader.readLine()) != null) {
+                String[] userDetails = line.split(" ");
+                String uname = userDetails[0];
+
+                if (username.equals(uname)) {
+                    return true;
+                }
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return false;
     }
 
     public void register() throws IOException {
@@ -57,9 +81,10 @@ public class UserRegistration {
         //Referred URL : https://www.baeldung.com/java-map-entry
         //Printing all our security questions(keys) and answers(value) in the file
         for (Map.Entry<Integer, String> entryAnswer : securityQuest.entrySet()) {
-            fileWriter.append((entryAnswer.getValue()));
+//            fileWriter.append((entryAnswer.getValue()));
+            System.out.println(entryAnswer.getValue());
             securityAnswers.add(in.nextLine());
         }
-
+        registerUser(username, password, securityAnswers);
     }
 }
